refactor(web): move avatar to base ui (#34889)
