+++ web/app/components/base/ui/avatar/index.stories.tsx
