+++ web/app/components/workflow/nodes/human-input/components/delivery-method/recipient/member-list.tsx
@@ -4,8 +4,8 @@ import type { Recipient } from '@/app/components/workflow/nodes/human-input/type
 import type { Member } from '@/models/common'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Avatar } from '@/app/components/base/avatar'
 import Input from '@/app/components/base/input'
+import { Avatar } from '@/app/components/base/ui/avatar'
 import { cn } from '@/utils/classnames'
 
 const i18nPrefix = 'nodes.humanInput'
